# apps/backend/demoYuno/src/main/java/com/example/demo/dto/ContinuePaymentRequest.java
package com.example.demo.dto;

public class ContinuePaymentRequest {
    private String paymentId;
    private String paymentMethodToken;

    public String getPaymentId() {
        return paymentId;
    }

    public void setPaymentId(String paymentId) {
        this.paymentId = paymentId;
    }

    public String getPaymentMethodToken() {
        return paymentMethodToken;
    }

    public void setPaymentMethodToken(String paymentMethodToken) {
        this.paymentMethodToken = paymentMethodToken;
    }
}