# apps/backend/demoYuno/src/main/java/com/example/demo/dto/StartPaymentRequest.java
package com.example.demo.dto;

public class StartPaymentRequest {
    private Double amount;
    private String currency;
    private String customerId;

    public Double getAmount() {
        return amount;
    }

    public void setAmount(Double amount) {
        this.amount = amount;
    }

    public String getCurrency() {
        return currency;
    }

    public void setCurrency(String currency) {
        this.currency = currency;
    }

    public String getCustomerId() {
        return customerId;
    }

    public void setCustomerId(String customerId) {
        this.customerId = customerId;
    }
}